+++ votes.py
@@ -9,6 +9,7 @@ class VoteHandler(base.BaseHandler):
         _event_id = self.get_int_argument('event_id')
         await self.require_event_access(_event_id)
         _suggestion_id = self.get_int_argument('id')
+        self.require_event_suggestion(_suggestion_id, _event_id)
         # check if vote exists
         c = self.db.execute_rowcount(
             "SELECT * FROM willbeout_votes WHERE user_id = %s and event_id = %s and suggestion_id = %s",
@@ -31,6 +32,7 @@ class ChangeVoteHandler(base.BaseHandler):
         _suggestion_id = self.get_int_argument('id')
         _event_id = self.get_int_argument('event_id')
         await self.require_event_access(_event_id)
+        self.require_event_suggestion(_suggestion_id, _event_id)
         _user_id = self.get_current_user()['id']
         self.db.execute(
             "DELETE FROM willbeout_votes WHERE suggestion_id = %s AND user_id = %s AND event_id = %s",